245. Getting the data from the confirmation modal

+++ DatingApp/client/src/app/messages/messages.component.ts
@@ -1,3 +1,4 @@
+import { ConfirmService } from './../services/confirm.service';
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../models/message';
 import { Pagination } from '../models/pagination';
@@ -16,9 +17,12 @@ export class MessagesComponent implements OnInit {
   pageNumber: number = 1;
   pageSize: number = 5;
 
-  loading= false;
+  loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(
+    private messageService: MessageService,
+    private confirmService: ConfirmService //1. inject the service
+  ) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -35,19 +39,24 @@ export class MessagesComponent implements OnInit {
   }
 
   pageChanged(event: any): void {
-    if(this.pageNumber !== event.page) {
+    if (this.pageNumber !== event.page) {
       this.pageNumber = event.page;
       this.loadMessages();
     }
   }
 
-  //1. add a method to delete a message
-  deleteMessage(id:number) {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
-    });
+  deleteMessage(id: number) {
+    //2. use the confirm service to show the modal, to insure the user is sure they want to delete the message
+    this.confirmService.confirm('Confirm Delete Message', 'This cannot be undone!').subscribe(result => {
+      if (result) {
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+        });
+      }
+    })
+    //3. back to README.md
+
   }
-  //2. go to the html to invoke this method
